refactor: bottomsheet 폴더 위치 변경 #26
