+++ src/app/concert/[id]/_components/bottom-sheet/bottom-sheet.tsx
@@ -3,8 +3,8 @@
 import Link from 'next/link';
 
 import styles from './bottom-sheet.module.scss';
-import NegativeButton from '../button/negative-button';
-import PositiveButton from '../button/positive-button';
+import NegativeButton from '../../../../../shared/components/button/negative-button';
+import PositiveButton from '../../../../../shared/components/button/positive-button';
 
 interface BottomSheetProps {
   onClose: () => void;
