+++ src/app/concert/[id]/page.tsx
@@ -6,7 +6,7 @@ import { useInfiniteQuery } from '@tanstack/react-query';
 import Link from 'next/link';
 import { useInView } from 'react-intersection-observer';
 
-import BottomSheet from '@/shared/components/bottom-sheet/bottom-sheet';
+import BottomSheet from '@/app/concert/[id]/_components/bottom-sheet/bottom-sheet';
 import AppBarSetter from '@/shared/components/header/app-bar/app-bar-setter';
 
 import UserCard from './_components/user-card/user-card';
